Software/backend/Services/Inference/Segmentation/requirements.txt
pika==1.3.2
dicomweb-client==0.59.1
python-dotenv==1.0.1
SimpleITK==2.3.1
pydicom-seg==0.4.1
